+++ src/widget/button.rs
@@ -184,7 +184,7 @@ impl Widget for Button {
         } else {
             frame.register_hitbox(self.agent_id.clone(), area, 1);
             if let Some(msg) = self.on_click.take() {
-                frame.register_message(self.agent_id.clone(), msg);
+                frame.register_message(self.agent_id.clone(), "click", msg);
             }
         }